requirements.txt
# Miko3 Talents Automation Framework - Dependencies
PyYAML>=6.0
Jinja2>=3.1.0
Pillow>=10.0.0
pytest>=7.0.0
colorama>=0.4.6
opencv-python>=4.8.0
scikit-image>=0.22.0
numpy>=1.24.0
gtts>=2.4.0
edge-tts>=6.1.0
pydub>=0.25.1
